チップの改変とプレイヤーの追加、physicsの追加

+++ source/utility/BackGround.cpp
@@ -52,7 +52,7 @@ void BackGround::Initialize( )
 int BackGround::clearData( int arg_bmpNo )
 {
 	setUseFlg( arg_bmpNo , false ) ;
-	setBmpXY( arg_bmpNo , 0 , 0 ) ;
+	setBmpXY( arg_bmpNo , 0.0f , 0.0f ) ;
 	setBmpUV( arg_bmpNo , 0 , 0 ) ;
 	setBmpWH( arg_bmpNo , 0 , 0 ) ;
 	setBmpScale( arg_bmpNo , 0 , 0 ) ;
@@ -112,7 +112,7 @@ int BackGround::setUseRotate( int arg_bmpNo , bool arg_useRotate )
 /*/
 /*	ポジションのセット
 /*/
-int BackGround::setBmpXY( int arg_bmpNo , int arg_x , int arg_y )
+int BackGround::setBmpXY( int arg_bmpNo , float arg_x , float arg_y )
 {
 	bmpBGTable_[ arg_bmpNo ]._x = arg_x ;
 	bmpBGTable_[ arg_bmpNo ]._y = arg_y ;
@@ -189,7 +189,7 @@ int BackGround::setAngle( int arg_bmpNo , float arg_degree )
 int BackGround::setBmpData(
 		int arg_bmpNo ,
 		int arg_anchor ,
-		int arg_x , int arg_y ,
+		float arg_x , float arg_y ,
 		int arg_u , int arg_v ,
 		int arg_w , int arg_h ,
 		float arg_scaleX , float arg_scaleY ,
